Feat: recursively render whole directory as one collection
The main way I like to structure my cards is to have it so that the
directory structure of my markdown files is the same structure as the
collection and all the (sub-)decks. I.e., the root directory name is the
root name of the collection and the name of the markdown file inside is
the name of the deck. This should manifest as a deck with the name
"<Root>::<FileName>"

Media files are automatically linked correctly and can be located
anywhere on a user's system.