Add (HorizontalRule) as a terminal character for extended card
Issue: Take the case where we mean to create a simple card (H1 until
next card) but that card is followed by an extended card beginning with
a horizontal rule. When we had a simple card followed by an extended
card, the `cardExtFrontBody` parser combinator wasn't failing when it
encountered a horizontal rule. This means everything until the next h1
or `. . .` would be included in the front of the simple card.

The grammar is currenty:

```
H1
many (block) except (ExtCardSep^*^ | H1)
...Back Of card
```

But should be:

```
H1
many (block) except (ExtCardSep^*^ | H1 | HorizontalRule)
```

* Para [Str ".", Space, Str ".", Space, Str ".", Space]