Feat: ignore files at each deck level
I wanted to have it so that I can have README files in the root of a
deck. But... these are markdown files. So they would be converted to
cards.

You can now use a special file in each directory of your collection
called ".pankyignore" to mark the files you don't want to be compiled.

For now, the extend of the functionality is just NAMEs *not* paths, can
be included in the .pankyignore file and the *only work for that
directory level*. The names can also be of directories. But these should
not be affixed with a slash.

If I think it's necessary, or other's demand it, there is scope to make
this file function kind of like a git ignore file. We would need to
decide if we should just have one ignore file at the root, or at every
level. Or, dare I say it, both.